LayoutTests/fast/forms/date/calendar-picker-with-step-expected.txt

Test if step limits available dates in calendar picker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Step without min.
PASS availableDatesInCurrentMonth().join(",") is ["2011-05-08", "2011-05-17", "2011-05-26"].join(",")
Step when min is set.
PASS availableDatesInCurrentMonth().join(",") is ["2011-05-01", "2011-05-10", "2011-05-19", "2011-05-28"].join(",")
 
